Creating the ServerSideViewSet class for Datatables
