+++ squalaetp/ajax.py
@@ -1,6 +1,4 @@
 from django.http import JsonResponse
-from rest_framework.response import Response
-from rest_framework import viewsets, permissions, status
 
 from .serializers import (
     XelonSerializer, XELON_COLUMN_LIST, XelonTemporarySerializer, XELON_TEMP_COLUMN_LIST,
@@ -8,29 +6,15 @@ from .serializers import (
 )
 from .models import Xelon, XelonTemporary, SparePart, Sivin
 
-from utils.django.datatables import QueryTableByArgs
+from utils.django.datatables import ServerSideViewSet
 from utils.django.validators import VIN_OLD_PSA_REGEX
 
 
-class XelonViewSet(viewsets.ModelViewSet):
-    permission_classes = (permissions.IsAuthenticated,)
+class XelonViewSet(ServerSideViewSet):
     queryset = Xelon.objects.filter(date_retour__isnull=False)
     serializer_class = XelonSerializer
-
-    def list(self, request, **kwargs):
-        try:
-            self._filter(request)
-            xelon = QueryTableByArgs(self.queryset, XELON_COLUMN_LIST, 1, **request.query_params).values()
-            serializer = self.serializer_class(xelon["items"], many=True)
-            data = {
-                "data": serializer.data,
-                "draw": xelon["draw"],
-                "recordsTotal": xelon["total"],
-                "recordsFiltered": xelon["count"],
-            }
-            return Response(data, status=status.HTTP_200_OK)
-        except Exception as err:
-            return Response(err, status=status.HTTP_404_NOT_FOUND)
+    column_list = XELON_COLUMN_LIST
+    column_start = 1
 
     def _filter(self, request):
         query = request.query_params.get('filter', None)
@@ -45,64 +29,24 @@ class XelonViewSet(viewsets.ModelViewSet):
             self.queryset = Xelon.search(query)
 
 
-class TemporaryViewSet(viewsets.ModelViewSet):
-    permission_classes = (permissions.IsAuthenticated,)
+class TemporaryViewSet(ServerSideViewSet):
     queryset = XelonTemporary.objects.all()
     serializer_class = XelonTemporarySerializer
-
-    def list(self, request, **kwargs):
-        try:
-            query = QueryTableByArgs(self.queryset, XELON_TEMP_COLUMN_LIST, 1, **request.query_params).values()
-            serializer = self.serializer_class(query["items"], many=True)
-            data = {
-                "data": serializer.data,
-                "draw": query["draw"],
-                "recordsTotal": query["total"],
-                "recordsFiltered": query["count"],
-            }
-            return Response(data, status=status.HTTP_200_OK)
-        except Exception as err:
-            return Response(err, status=status.HTTP_404_NOT_FOUND)
+    column_list = XELON_TEMP_COLUMN_LIST
+    column_start = 1
 
 
-class StockViewSet(viewsets.ModelViewSet):
-    permission_classes = (permissions.IsAuthenticated,)
+class StockViewSet(ServerSideViewSet):
     queryset = SparePart.objects.all()
     serializer_class = SparePartSerializer
-
-    def list(self, request, **kwargs):
-        try:
-            query = QueryTableByArgs(self.queryset, SPAREPART_COLUMN_LIST, 0, **request.query_params).values()
-            serializer = self.serializer_class(query["items"], many=True)
-            data = {
-                "data": serializer.data,
-                "draw": query["draw"],
-                "recordsTotal": query["total"],
-                "recordsFiltered": query["count"],
-            }
-            return Response(data, status=status.HTTP_200_OK)
-        except Exception as err:
-            return Response(err, status=status.HTTP_404_NOT_FOUND)
+    column_list = SPAREPART_COLUMN_LIST
 
 
-class SivinViewSet(viewsets.ModelViewSet):
-    permission_classes = (permissions.IsAuthenticated,)
+class SivinViewSet(ServerSideViewSet):
     queryset = Sivin.objects.all()
     serializer_class = SivinSerializer
-
-    def list(self, request, **kwargs):
-        try:
-            sivin = QueryTableByArgs(self.queryset, SIVIN_COLUMN_LIST, 1, **request.query_params).values()
-            serializer = self.serializer_class(sivin["items"], many=True)
-            data = {
-                "data": serializer.data,
-                "draw": sivin["draw"],
-                "recordsTotal": sivin["total"],
-                "recordsFiltered": sivin["count"],
-            }
-            return Response(data, status=status.HTTP_200_OK)
-        except Exception as err:
-            return Response(err, status=status.HTTP_404_NOT_FOUND)
+    column_list = SIVIN_COLUMN_LIST
+    column_start = 1
 
 
 def xelon_prod_async(request):
